backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose==3.3.0
pydantic==2.10.3
pydantic-settings==2.6.1
